Convert to TypeScript
& Directory structure optimization
