--- main.js
@@ -1,28 +1,0 @@
-const { app, BrowserWindow } = require('electron')
-const path = require('node:path')
-
-const createWindow = () => {
-    const win = new BrowserWindow({
-        width: 800,
-        height: 600,
-        webPreferences: {
-            preload: path.join(__dirname, 'preload.js')
-        }
-    })
-  
-    win.loadFile('index.html')
-}
-
-app.whenReady().then(() => {
-    createWindow()
-
-    // Open a new window when activating & there is no window (macOS)
-    app.on('activate', () => {
-        if (BrowserWindow.getAllWindows().length === 0) createWindow()
-      })
-})
-
-// Quit app when all windows are closed (Windows & Linux)
-app.on('window-all-closed', () => {
-    if (process.platform !== 'darwin') app.quit()
-  })
