+++ src/renderer/renderer.ts
@@ -1,0 +1,1 @@
+console.log('Hello from renderer!');
